chore: restructure web app
