+++ web/public/assets/qr-bg.jpg
